Syntax: Don't highlight fields (#8)
* syntax: don't try to highlight fields

Fields are annoying, because there's no way to know what kind of object
they really are (short of implementing full type checking, which is not
feasible in a little extension like this). For example, in
`$myVar.myField`, `myField` could be a normal field or a dictionary key.

We feel it's safer not to highlight than to highlight as any particular
data type and risk giving false information. Also, we don't want to make
the whole screen a mass of random colors -- leaving some text
un-highlighted is a good thing.

* syntax: explain complicated highlighting regex