functions: yank buf to system clip command (#9)
Adds the YagCopy() function and :YagCopy ex command (which just calls
the function). The actual command used is the ex-mode `%y *`, which
yanks the entire buffer contents into the `"*` buffer. This buffer is
linked to the system clipboard, so the user can then paste the CC into
the YAGPDB Control Panel.

If the user's vim was not compiled with clipboard support, a warning
message is echoed instead.