test[syntax]: begin adding syntax matching tests
There are still a few "pass if 1 equals 1" stubs in here, marked by
TODOs. I'll get to them later.

test: apparently colons aren't allowed in descriptions

test: add missing colon :(

test: remove leading spaces on comments

test/syntax: setup env properly

test: adjust "correct answers" for now

The fact that I have to do this means there are some pretty serious
issues with the highlighting logic, and I definitely need to fix those
later. To be fair, I kind of expected this, but it's more serious now.

That's the point of tests, though!

[skip ci] - there's another commit coming soon with test readability
improvements

test: switch order for test statements

This change greatly improves the readability of error messages generated
during the testing process.