ci: implement test cases (#11)
* ci: initial commit - add automated pipeline

ci: actually source the plugin files

ignore build artifacts

* test[syntax]: begin adding syntax matching tests

There are still a few "pass if 1 equals 1" stubs in here, marked by
TODOs. I'll get to them later.

test: apparently colons aren't allowed in descriptions

test: add missing colon :(

test: remove leading spaces on comments

test/syntax: setup env properly

test: adjust "correct answers" for now

The fact that I have to do this means there are some pretty serious
issues with the highlighting logic, and I definitely need to fix those
later. To be fair, I kind of expected this, but it's more serious now.

That's the point of tests, though!

[skip ci] - there's another commit coming soon with test readability
improvements

test: switch order for test statements

This change greatly improves the readability of error messages generated
during the testing process.

* github: update CODEOWNERS

Reflect the work that went into the files and credit the appropriate author.
Also, sort the entries for readability.

* syntax: rework syntax regions

This commit does away with the infernally annoying `yagpdbccIgnore`
syntax region, which was a previous attempt to make normal syntax
highlighting not apply to regular text. This is replaced with proper Vim
highlight syntax, mostly consisting of tagging syntax items with the
`contained` keyword.

* tests: set indentation format

Vader.vim requires two-space indendentation.
We achieve a consistent setting by using the modline.

* all: declare vader.vim as dev-dependency

So that tests can be run locally, vader.vim is required.
This makes it a dev-dependency, and as such, must be mentioned.

* ci: add clean target

* github: fix CODEOWNER formatting

Co-authored-by: Lucas Ritzdorf <42657792+LRitzdorf@users.noreply.github.com>