ci: set up vint (#17)
* ci: set up vint

As discussed, project maintainers agreed on setting up a linting tool
for the Vim script programming language. This commit adds said tool with
a configured GitHub workflow as well as a basic configuration file.

Closes #16

ci: use JSON output

Hopefully this'll make it comment on files /shrug

test workflow stuff

ci: fix it yay

* ci: clean up code, exit 1 on lint errors

* ci: don't fail if only style errors exist

* ci/lint: don't install Vim

Not used, thus not needed to be installed.

* ci/lint: restructure vint.py

Implement feedback and give the whole file some polish.

Co-Authored-By: Joe <56809242+Jo3-L@users.noreply.github.com>

* ci/lint: remove actions/setup-python

As Python 3.8.10 comes preinstalled on `ubuntu-latest` runners, this
step adds unnecessary overhead and confusion, thus this commit removes
said step in the workflow configuration.

* vint: don't use shell=True with Popen

Though it's not really a concern here, this is a best practice. Misusing
shell=True can, in some cases, allow for command injection
vulnerabilities (usually with user input, which again, isn't a concern
for us at this point).

Co-authored-by: Joe <56809242+Jo3-L@users.noreply.github.com>
Co-authored-by: Lucas Ritzdorf <42657792+LRitzdorf@users.noreply.github.com>