functions/clipboard: config var for clip register
With this commit, a new configuration variable `g:yagpdb_use_primary` is
introduced, which determines whether to write to the `"*` (PRIMARY) or
`"+` CLIPBOARD register.

The difference in these is, that PRIMARY uses the middle mouse button to
insert its content, whereas CLIPBOARD uses Ctrl + V -- thus, we will
from now on default to writing to `"+`, as that is probably what most
users are used to and expect to happen.

Documentation of this variable has been added to the README file.

Closes #21