feature: implement code completion (#24)
* initial work on code-completion

* completion: further progress on code generation

This commit adds parsing all files and generating a list of completion
labels, which are as of now printed to stdout. Further commits will base
on this and finally, allow us to fully generate these sources.

* completion: first generation of sources as final step

Currently the sources are not loaded and throw errors when loading them.
Future commits will have to look into that. Currently it is suspected
that https://github.com/hrsh7th/nvim-cmp gives wrong or incomplete
instructions.

* completion: add instructions and license header

This commit concludes the development of this feature by adding instructions on how to
enable code completion via the user's neovim configuration.

It also adds a license header to all new files which haven't received one yet.

* completion: use correct source name

* completion: extract more boilerplate to template file

* completion: be more pythonic

Remove unnecessary code and write it a bit more efficiently.

* completion: be a bit more compact

* completion/workflows: automatically regenerate sources

* completion: be only available when ft is yagpdbcc

* completion: use correct path-matcher syntax