tests: update YagCopy to trim extra trailing newline
When copying an entire buffer with `:%y` (as the YagCopy command does),
Vim includes the trailing newline at the end of the file. This isn't
really necessary when the command code is pasted into the YAGPDB
dashboard. Also, this extra newline isn't even rendered in Vim, but it
is still taken into account when pasting text (which was messing up our
unit tests, too).