tests: remove YagCopy clipboard test
The YagCopy test (for the clipboard feature) has been a bit of a pain to
get working. First, it appears that GitHub's runners don't have
clipboard support available, which makes our CI test fail, even if it
would pass locally. Also, we'd ideally just skip this test if clipboard
support isn't present, but that also proves to be rather difficult.

We may consider re-adding a similar test in the future, but for now,
it's more trouble than it's worth.