Fix: YagCopy command and related tests (#31)
* tests: update YagCopy test to clear the buffer

We now clear the buffer before pasting in the code we just copied. This
allows us to properly verify that the copied text is what we expect.

* tests: update YagCopy to trim extra trailing newline

When copying an entire buffer with `:%y` (as the YagCopy command does),
Vim includes the trailing newline at the end of the file. This isn't
really necessary when the command code is pasted into the YAGPDB
dashboard. Also, this extra newline isn't even rendered in Vim, but it
is still taken into account when pasting text (which was messing up our
unit tests, too).

* test: only execute YagCopy when clipboard present

* workflows: also install neovim

workflows: use correct package and make target

workflows: use correct command to execute

hmmmmmmm

* Makefile: small indentation fix

* tests: remove YagCopy clipboard test

The YagCopy test (for the clipboard feature) has been a bit of a pain to
get working. First, it appears that GitHub's runners don't have
clipboard support available, which makes our CI test fail, even if it
would pass locally. Also, we'd ideally just skip this test if clipboard
support isn't present, but that also proves to be rather difficult.

We may consider re-adding a similar test in the future, but for now,
it's more trouble than it's worth.

Co-authored-by: Luca Z <l-zeuch@email.de>