ftdetect: don't use augroups in ftdetect
This commit modifies `ftdetect/yagpdbcc.vim` to not use augroups.
The linter warning for autocommands without a group has been disabled as
well.

From `:h ftdetect`:
Note that there is no "augroup" command, this has already been done when
sourcing your file. You could also use the pattern "*" and then check
the contents of the file to recognize it.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>