workflows: use GH-CLI to checkout and push to PR (#42)
* workflows: use GH-CLI to checkout and push to PR

This commit modifies the generate.yml workflow to use the pre-installed
GitHub CLI which comes on all GitHub-hosted runners. It is used to
properly checkout the pull request, which then allows us to actually
push automagic regeneration of completion sources.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* workflows: consistent name cases

Be consistent with naming workflow steps. First letter is capitalised, the rest lowercase except for abbreviations.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>