refactor: optimise checks for global variables
This commit optimises the test for the documented global configuration
variables.
From `:h get()`:
get({list}, {idx}, [, {default}])
    Get item {idx} from {list}. When item is not available, return
    {default}. Return zero when {default} is omitted.

The benefit of this function is first and foremost readability, as it
only requires one clean call instead of chaining two things together.
Secondly, it is more intuitive to read, as well as easier to maintain.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>