all: refactor codebase for autoload capabilities
This commit touches up the whole codebase to add basic autoloading
capabilities to this plugin. It also replaces the plugin/yagpdbcc.lua
file with plugin/yagpdbcc.vim, for better compatibility reasons.

This allows us to manipulate both Vim and Neovim to our liking.
The override_ft functionality has been moved to the aforementioned new
file, in hopes that it will work better there; see the issue that the
config variable has to be set before loading, as per the README.

For *reasons*, the note will remain there until we can verify independly
that this is indeed an improvement over the current status quo.

Future work entails smaller refactor commits to make the overall
software more maintainable, as well as extensible, whenever appropriate.
We should also consider getting around managing folds properly, as of
now that capability still is missing.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>