ci: test with newer Vim version
The Vim version provided by GitHub hosted runners is faulty in the sense
that it fails the test suite, despite all being successful.

This commit modifies the workflows/ci.yml file to test with the latest
available version on ArchLinux, which would be v8.2.5046, as that is a
known working version.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>