all: refactor codebase for autoload capabilities (#49)
* all: refactor codebase for autoload capabilities

This commit touches up the whole codebase to add basic autoloading
capabilities to this plugin. It also replaces the plugin/yagpdbcc.lua
file with plugin/yagpdbcc.vim, for better compatibility reasons.

This allows us to manipulate both Vim and Neovim to our liking.
The override_ft functionality has been moved to the aforementioned new
file, in hopes that it will work better there; see the issue that the
config variable has to be set before loading, as per the README.

For *reasons*, the note will remain there until we can verify independly
that this is indeed an improvement over the current status quo.

Future work entails smaller refactor commits to make the overall
software more maintainable, as well as extensible, whenever appropriate.
We should also consider getting around managing folds properly, as of
now that capability still is missing.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* completion: regenerate sources

* autoload: move config functions to appropriate location

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* ci: test with newer Vim version

The Vim version provided by GitHub hosted runners is faulty in the sense
that it fails the test suite, despite all being successful.

This commit modifies the workflows/ci.yml file to test with the latest
available version on ArchLinux, which would be v8.2.5046, as that is a
known working version.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* plugin: move YagCopy(), move command registration

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* readme: remove config note for ft override

With the new autoloading functionality, this note is no longer
necessary, as it does not matter where the documented configuration
variable needs to be set.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* repo: update CODEOWNERS

In light of the new additions to the repository, the CODEOWNERS file
needs a small update as well.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* fixup! repo: update CODEOWNERS

* plugin: override_ft: make autocmds obey current override setting

With this change, we no longer depend on setting `g:yagpdbcc_override_ft` on
startup. Autocommands will now check this value at runtime, so mid-session
changes are respected now.

* filetypes: move new conditional autocmds to ftdetect

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Lucas Ritzdorf <42657792+LRitzdorf@users.noreply.github.com>