ci: use a matrix strategy, sep jobs, add linters
This commit modifies the ci.yml file to use a matrix strategy to test across
multiple versions, both for Neovim and Vim. Secondly, it moves the former
test-nvim step to a separate job, in hopes that the pipeline will complete
faster.

The second part of this commit includes the new lint.yml file, which now
contains the former vim-lint job, as well as two new jobs; Linting via Pyright
for Python, and markdown via Markdownlint.

Future work might entail caching the dependencies on Vim and Neovim, so that we
can speed up the pipeline even further.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>