tests: add more test cases (#65)
* test: address first TODO in syntax.vader

This commit addresses the first TODO in the syntax.vader test file, testing the
code snippet with a variety of syntax elements.

At this stage, one test is failing: we're not matching format verbs inside raw
strings, though they definitely work there and as such, should be highlighted.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* syntax: add yagFormat to yagRawStr group

This commit adds the yagFormat group to the yagRawStr region.
Without this change, we would not match format verbs inside raw strings,
thus creating an antipattern.
Not highlighting these verbs might cause the user
to think that format verbs are not supported (do not work) inside raw strings,
while -- in fact -- they do work.

This change also fixes the previously failing test suite.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: address last TODO in syntax.vader

This commits addresses the second TODO in the syntax.vader test file, checking
the code snippet with a lot of nested syntax.

All tests are passing locally.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: add test suite for autoloaded functions

This commit adds a small test suite for a few autoloaded functions.

As of this commit, that suite still fails but the following commit will fix
that.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* autoload/config: default to empty string

This commit modifies the yagpdbcc#config#SnippetEngine() function to return an
empty string if the variable is not set. It is documented as a string config
value, thus the default value should be an empty string.

This commit also fixes the previously failing test suite.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: split autoload contexts

Currently, tests fail for vim 8.0.0000. In hopes that it is some sort of
name-clash for multiple variables as indicated by the logs, we'll split the
files accordingly.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* autoload: remove `unlet` directive

Apparently, Vim 8.0.0000 does not know how to work with an `unlet`ed variable.
See also: thinca/vim-quickrun#133

Signed-off-by: Luca Zeuch <l-zeuch@email.de>