test: write tests for indentation
This commit adds a new test suite for indentation. Fixups for the indent script
itself will follow.

The vimrc had to be amended with the following two lines to make it work with
our vader files:

```diff
+ set shiftwidth=2
+ set expandtab
```

This is given by how vader files are structured with their two-space based
indentation designating the various blocks.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>