indent: add indentation logic (#68)
* feat: add indentation logic

This commit adds an indentexpr to tell Vim how to properly indent YAGPDB-CC
code.
The script is based on fatih/vim-go's gohtmltmpl.vim indentexpr, which YAGPDB
custom commands essentially are. Proper credit has been given.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* github: amend CODEOWNERS with indent

Though this is mostly adapted, a proper maintainer should still be defined
within the repository.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* indent: return -1 as default

As per `:h 'indentexpr'`, if the set indentexpr returns -1, it will keep the
current indent. This is favourable over a 0 default value, as the latter
indicates "zero levels of indentation", i.e. pos 0 of the whole line.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: write tests for indentation

This commit adds a new test suite for indentation. Fixups for the indent script
itself will follow.

The vimrc had to be amended with the following two lines to make it work with
our vader files:

```diff
+ set shiftwidth=2
+ set expandtab
```

This is given by how vader files are structured with their two-space based
indentation designating the various blocks.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* indent: fixes following test suite

This commit summarises the changes mandated by the previously added testsuite
for indenting YAGPDB custom command code.

They are all passing locally.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

Signed-off-by: Luca Zeuch <l-zeuch@email.de>