scripts/vint: exclude `.bundle` from linter
With the current implementation, developers who keep the `.bundle` directory
around and run `make lint`, vint will also be invoked on that directory.

This behaviour is undesirable and confusing -- thus, this commit changes the
construction of the command to first glob the project tree for any Vim Script
files. Given how globbing a path ignores dotfiles, this was quite
straightforward.

The only downside is that now the command is *very* long, however that should
not be an issue -- in fact, I believe it is better that way, because it becomes
immediately obvious which files were linted, even when no violations were found.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>