test/autoload/util: add util test suite
The old test for yagpdbcc#util#PathSep() is copied directly, it works
well.

Add two new tests for yagpdbcc#util#Error() and yagpdbcc#util#Warn()
respectively, though these functions are as of now (nigh) impossible to
test. A TODO was added to reflect that sentiment.

Perhaps in a future patch we should consider logging it to `:messages`
via `echom` instead of `echo`, and then perform a search on `:messages`.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>