test/syntax/identifiers: add tests for identifiers
Add a decently sized test suite for all identifier-variants I could think of.

With the current implementation of `yagIdentifier`,
we are not good at not marking invalid ones, such as `$.` or `$kebab-case`.
In YAGPDB, these are invalid, though we are happily highlighting them,
indicating that we think they are valid.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>