test: spec out new structure (#70)
* test: document test directory structure

This commit drafts out a new structure for our tests. This is up for review and
changes, fixups will follow as-needed.

Please refer to the patch of this commit to review the new draft.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/doc: small grammatical corrections

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* fixup! test/doc: small grammatical corrections

* fixup! test: document test directory structure

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: remove all old tests, add new syntax test

Add a newly rewritten syntax tests, specifically for comments only.

Further remove all currently present tests, for easier working the base up and
gradual testing.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax/comments: restore after running suite

Not restoring the prior state may have undesirable side effects, so it is better
to prevent them, even if they will not occur.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax/numbers: add tests for numbers

This commit adds a whole suite dedicated to all kinds of numbers.
It is rather minimal, as there aren't many combinations that we have to test
for.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax: increment year of copyright

This PR is taking rather long to work on, so updating it to the year where it is
shipped may be advisable.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax: do not test keywords and identifiers

Testing for things that are not actually scope of the suite are an antipattern
and will only cause confusion in the long run.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* tests: add indentation tests

Add a test suite for indentation. Given how the old tests are sufficient, this
commit just adds that file back to the worktree.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/autoload/config: add autoloaded config func test suite

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/readme: remove mention of main test file

Remove `test.vader` bullet point from README.md, because it isn't
needed. Instead, vader takes care of running each found vader file
individually.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: move indent.vader to feature/

According to the test README, single test files are to be placed in
feature/, if the tested feature is not too complex.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/autoload/util: add util test suite

The old test for yagpdbcc#util#PathSep() is copied directly, it works
well.

Add two new tests for yagpdbcc#util#Error() and yagpdbcc#util#Warn()
respectively, though these functions are as of now (nigh) impossible to
test. A TODO was added to reflect that sentiment.

Perhaps in a future patch we should consider logging it to `:messages`
via `echom` instead of `echo`, and then perform a search on `:messages`.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax/comments: improve feedback

Improve feedback from the assertion that a variable inside a comment should be
matched as a comment, as the wanted result wasn't very well formulated.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax/identifiers: add tests for identifiers

Add a decently sized test suite for all identifier-variants I could think of.

With the current implementation of `yagIdentifier`,
we are not good at not marking invalid ones, such as `$.` or `$kebab-case`.
In YAGPDB, these are invalid, though we are happily highlighting them,
indicating that we think they are valid.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test: improve colour output for TODO note

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* test/syntax: add string test suite

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>