workflows: general housekeeping
There's now Neovim 0.10.0, so add the version prior to that to the version
matrix.

Rename the workflow files genfuncs and generate to be a bit more telling about
what they contain.

Tell lytfs about our GITHUB_TOKEN, thereby authenticating with the GitHub API.
Although the unauthenticated ratelimit is quite lenient considering that
Workflows take a bit to fire and run, it's better to be safe, especially when
quick-firing multiple runs.

Checkout the repository with an SSH deploy key, in hopes to cause less headaches
when trying to push regenerated syntax defs and suchlike to a PR.

Attempt to cache the build Neovim and Vim binaries, hopefully speeding up
workflow runs. Not quite sure if it will work, but worst case scenario is
actions/cache emitting a warning, as far as I can tell.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>