repo: general housekeeping (#85)
* github: fix broken link to contrib guidelines

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* build: rename generate -> completion

Rename the old generate target to "completion", thus being more clear about what
it does.

In a similar vein, we introduce a new meta-target "generate", that runs both
syntax and completion targets.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* workflows: general housekeeping

There's now Neovim 0.10.0, so add the version prior to that to the version
matrix.

Rename the workflow files genfuncs and generate to be a bit more telling about
what they contain.

Tell lytfs about our GITHUB_TOKEN, thereby authenticating with the GitHub API.
Although the unauthenticated ratelimit is quite lenient considering that
Workflows take a bit to fire and run, it's better to be safe, especially when
quick-firing multiple runs.

Checkout the repository with an SSH deploy key, in hopes to cause less headaches
when trying to push regenerated syntax defs and suchlike to a PR.

Attempt to cache the build Neovim and Vim binaries, hopefully speeding up
workflow runs. Not quite sure if it will work, but worst case scenario is
actions/cache emitting a warning, as far as I can tell.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* workflows: give a key to actions/cache

actions/cache wants a key to store the cache under. Add one.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* workflows: remove caching

It appears as if the setup-vim action is not quite compatible with caching, or
we do not fully understand how it functions yet.

For now, we will remove the cache attempt and just cope with the longer build
times.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

* workflows: run make generate instead of syntax

Run the `generate` target instead of the `syntax` target to lessen the noise in
the git log, though we squash it anyway.

Signed-off-by: Luca Zeuch <l-zeuch@email.de>

---------

Signed-off-by: Luca Zeuch <l-zeuch@email.de>